seeding data

+++ src/migration/1702886725237-delete-photo.ts
@@ -1,21 +1,42 @@
 import {
   MigrationInterface,
   QueryRunner,
+  Table,
   TableColumn,
   TableForeignKey,
 } from 'typeorm';
 
 export class DeletePhoto1702886725237 implements MigrationInterface {
   public async up(queryRunner: QueryRunner): Promise<void> {
-    const table = await queryRunner.getTable('photo');
-    const foreignKey = table.foreignKeys.find(
-      (fk) => fk.columnNames.indexOf('userId') !== -1,
-    );
-    await queryRunner.dropForeignKey('photo', foreignKey);
-    await queryRunner.dropColumn('photo', 'userId');
+    if (await queryRunner.hasTable('photo')) {
+      const table = await queryRunner.getTable('photo');
+      const foreignKey = table.foreignKeys.find(
+        (fk) => fk.columnNames.indexOf('userId') !== -1,
+      );
+      await queryRunner.dropForeignKey('photo', foreignKey);
+      await queryRunner.dropColumn('photo', 'userId');
+      await queryRunner.dropTable('photo');
+    }
   }
 
   public async down(queryRunner: QueryRunner): Promise<void> {
+    await queryRunner.createTable(
+      new Table({
+        name: 'photo',
+        columns: [
+          {
+            name: 'id',
+            type: 'int',
+            isPrimary: true,
+          },
+          {
+            name: 'url',
+            type: 'varchar',
+          },
+        ],
+      }),
+      true,
+    );
     await queryRunner.addColumn(
       'photo',
       new TableColumn({
